refactor(stories): separate protocol contracts and markdown
